@@ -3,7 +3,7 @@ description: compralotodo.
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 1.0.43+2
+version: 1.0.44+2
 
 environment:
   sdk: ">=2.7.0 <3.0.0"
@@ -11,44 +11,48 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  auto_size_text: ^2.1.0
-  cached_network_image: ^2.2.0+1
-  camera: ^0.5.8+11
-  contacts_service: ^0.4.6
-  cupertino_icons: ^0.1.3
-  dio: ^3.0.9
-  flutter_cache_manager: ^1.2.0
+  auto_size_text: ^3.0.0
+  cached_network_image: ^3.1.0
+  camera:
+    git:
+      url: https://github.com/flutter/plugins
+      path: packages/camera/camera
+      ref: 9e46048ad2e1f085c1e8f6c77391fa52025e681f
+  contacts_service: ^0.6.3
+  cupertino_icons: ^1.0.4
+  dio: ^4.0.6
+  flutter_cache_manager: ^3.3.0
   flutter_emoji: ^2.2.1+1
-  flutter_local_notifications: ^1.4.4+2
+  flutter_local_notifications: ^9.5.2
   flutter_localizations:
     sdk: flutter
-  flutter_markdown: ^0.4.1
+  flutter_markdown: ^0.6.10
   flutter_masked_text: ^0.8.0
-  flutter_ringtone_player: ^2.0.0
-  fluttertoast: ^7.1.8
-  firebase_messaging: ^7.0.3
-  http: ^0.12.2
-  image_picker: ^0.6.7+12
+  flutter_ringtone_player: ^3.1.1
+  fluttertoast: ^8.0.9
+  firebase_core: ^1.10.0
+  firebase_messaging: ^11.1.0
+  http: ^0.13.3
+  image_picker: ^0.8.4+1
   image_cropper: ^1.3.1
   numeric_keyboard: ^1.0.0
-  searchable_dropdown: ^1.1.3
-  shared_preferences: ^0.5.7+3
-  package_info: ^0.4.3
-  path_provider: ^1.6.22
-  permission_handler: ^5.0.0
-  provider: ^4.3.2+2
-  pusher_websocket_flutter: ^0.2.0
-  share: ^0.6.5+4
-  sqflite: ^1.1.6
-  toast: ^0.1.5
-  url_launcher: ^5.5.2
-  webview_flutter: ^1.0.7
+  search_choices : ^2.0.19
+  shared_preferences: ^2.0.7
+  package_info: ^2.0.2
+  path_provider: ^2.0.4
+  permission_handler: ^8.1.6
+  provider: ^6.0.2
+  share: ^2.0.4
+  sqflite: ^2.0.2
+  url_launcher: ^6.0.10
+  web_socket_channel: ^2.2.0
+  webview_flutter: ^2.0.13
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.7.5
+  flutter_launcher_icons: ^0.9.1
 
 flutter:
   uses-material-design: true
