@@ -44,7 +44,7 @@ class _LoginPageState extends State<LoginPage> {
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: <Widget>[
                     Image(
-                      image: AssetImage("assets/logo/logo.png"),
+                      image: AssetImage("assets/logo/logo2.png"),
                       width: size.width/2,
                     ),
                     formLogin(), //form login
@@ -170,7 +170,7 @@ class _LoginPageState extends State<LoginPage> {
           ),
           Padding(
             padding: const EdgeInsets.fromLTRB(0.0, 15.0, 0.0, 15.0),
-            child: FlatButton(
+            child: TextButton(
               onPressed: () {
                 setState(() {
                   _passwordController.clear();
@@ -315,7 +315,7 @@ class _LoginPageState extends State<LoginPage> {
         if (result.isNotEmpty && result[0].rawAddress.isNotEmpty) {
 
           response = await http.post(
-            urlApi+"login",
+            Uri.parse(urlApi+"login"),
             headers:{
               'Accept': 'application/json',
               'Content-Type': 'application/json',
